Move ES sync package into search
